Rename base_snpe-> snpe_base (consistency).
