+++ sbi/inference/snpe/__init__.py
@@ -1,4 +1,4 @@
-from sbi.inference.snpe.base_snpe import SnpeBase
 from sbi.inference.snpe.snpe_a import SnpeA
 from sbi.inference.snpe.snpe_b import SnpeB
+from sbi.inference.snpe.snpe_base import SnpeBase
 from sbi.inference.snpe.snpe_c import SnpeC
