+++ sbi/inference/snpe/snpe_a.py
@@ -1,9 +1,11 @@
-import sbi.utils as utils
-from sbi.inference.snpe.base_snpe import SnpeBase
-import torch
 import os
+
+import torch
 from torch import distributions
 from torch.utils.tensorboard import SummaryWriter
+
+from sbi.inference.snpe.snpe_base import SnpeBase
+import sbi.utils as utils
 
 
 class SnpeA(SnpeBase):
@@ -22,7 +24,7 @@ class SnpeA(SnpeBase):
         prior,
         true_observation,
         num_pilot_samples=100,
-        density_estimator='maf',
+        density_estimator="maf",
         use_combined_loss=False,
         z_score_obs=True,
         retrain_from_scratch_each_round=False,
@@ -31,7 +33,7 @@ class SnpeA(SnpeBase):
         device=None,
     ):
         """
-        See base_snpe for docstring.
+        See snpe_base.SnpeBase for docstring.
 
         Args:
             num_atoms: int
@@ -39,17 +41,18 @@ class SnpeA(SnpeBase):
                 If -1, use all other parameters in minibatch.
         """
 
-        super(SnpeA, self).__init__(simulator=simulator,
-                                      prior=prior,
-                                      true_observation=true_observation,
-                                      num_pilot_samples=num_pilot_samples,
-                                      density_estimator=density_estimator,
-                                      use_combined_loss=use_combined_loss,
-                                      z_score_obs=z_score_obs,
-                                      retrain_from_scratch_each_round=retrain_from_scratch_each_round,
-                                      discard_prior_samples=discard_prior_samples,
-                                      device=device,
-                                      )
+        super(SnpeA, self).__init__(
+            simulator=simulator,
+            prior=prior,
+            true_observation=true_observation,
+            num_pilot_samples=num_pilot_samples,
+            density_estimator=density_estimator,
+            use_combined_loss=use_combined_loss,
+            z_score_obs=z_score_obs,
+            retrain_from_scratch_each_round=retrain_from_scratch_each_round,
+            discard_prior_samples=discard_prior_samples,
+            device=device,
+        )
 
         # Each APT run has an associated log directory for TensorBoard output.
         if summary_writer is None:
@@ -61,8 +64,6 @@ class SnpeA(SnpeBase):
             self._summary_writer = summary_writer
 
         raise NotImplementedError
-
-
 
     def _get_log_prob_proposal_posterior(self, inputs, context, masks):
         """
@@ -84,9 +85,7 @@ class SnpeA(SnpeBase):
 
         """
 
-        log_prob_posterior_non_atomic = self._neural_posterior.log_prob(
-            inputs, context
-        )
+        log_prob_posterior_non_atomic = self._neural_posterior.log_prob(inputs, context)
 
         batch_size = inputs.shape[0]
 
@@ -102,13 +101,11 @@ class SnpeA(SnpeBase):
         # of the parameters in the batch.
         assert 0 < num_atoms - 1 < batch_size
         probs = (
-                (1 / (batch_size - 1))
-                * torch.ones(batch_size, batch_size)
-                * (1 - torch.eye(batch_size))
+            (1 / (batch_size - 1))
+            * torch.ones(batch_size, batch_size)
+            * (1 - torch.eye(batch_size))
         )
-        choices = torch.multinomial(
-            probs, num_samples=num_atoms - 1, replacement=False
-        )
+        choices = torch.multinomial(probs, num_samples=num_atoms - 1, replacement=False)
         contrasting_inputs = inputs[choices]
 
         # We can now create our sets of atoms from the contrasting parameter sets
@@ -136,14 +133,12 @@ class SnpeA(SnpeBase):
         assert utils.notinfnotnan(log_prob_prior), "NaN/inf detected in prior eval."
 
         # Compute unnormalized proposal posterior.
-        unnormalized_log_prob_proposal_posterior = (
-                log_prob_posterior - log_prob_prior
-        )
+        unnormalized_log_prob_proposal_posterior = log_prob_posterior - log_prob_prior
 
         # Normalize proposal posterior across discrete set of atoms.
         log_prob_proposal_posterior = unnormalized_log_prob_proposal_posterior[
-                                      :, 0
-                                      ] - torch.logsumexp(unnormalized_log_prob_proposal_posterior, dim=-1)
+            :, 0
+        ] - torch.logsumexp(unnormalized_log_prob_proposal_posterior, dim=-1)
         assert utils.notinfnotnan(
             log_prob_proposal_posterior
         ), "NaN/inf detected in proposal posterior eval."
@@ -152,8 +147,5 @@ class SnpeA(SnpeBase):
             masks = masks.reshape(-1)
 
             log_prob_proposal_posterior = (
-                    masks * log_prob_posterior_non_atomic + log_prob_proposal_posterior
+                masks * log_prob_posterior_non_atomic + log_prob_proposal_posterior
             )
-
-        return log_prob_proposal_posterior
-
